+++ api/src/main/java/com/capstone/insurance/dto/policy/CustomerPolicyDto.java
@@ -5,6 +5,7 @@ import com.capstone.insurance.entities.enums.PolicyType;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -16,7 +17,7 @@ public class CustomerPolicyDto {
     private String policyCode;
     private String policyNumber; // The assigned policy number from CustomerPolicy
     private PolicyType policyType;
-    private Double coverageAmount;
+    private BigDecimal coverageAmount;
     private LocalDate startDate;
     private LocalDate endDate;
     private PolicyStatus status;
