+++ api/src/main/java/com/capstone/insurance/dto/policy/PolicyDto.java
@@ -5,6 +5,7 @@ import com.capstone.insurance.entities.enums.PolicyType;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -14,7 +15,7 @@ public class PolicyDto {
     private java.util.UUID id;
     private String policyCode;
     private PolicyType policyType;
-    private Double coverageAmount;
+    private BigDecimal coverageAmount;
     private LocalDate startDate;
     private LocalDate endDate;
     private PolicyStatus status;
